% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_methods.R
\name{dcMethods}
\alias{dcMethods}
\title{Get names of differential co-expression methods}
\usage{
dcMethods()
}
\value{
names of methods implemented
}
\description{
Returns a list of differential co-expression methods
}
\examples{
dcMethods()
}
