% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{dcNetwork}
\alias{dcNetwork}
\title{Generate a differential network from a DC analysis}
\usage{
dcNetwork(dcscores, dcpvals, thresh = NA)
}
\arguments{
\item{dcpvals}{a matrix, raw or adjusted p-values resulting from
\code{dcTest} or \code{dcAdjust} respectively}

\item{thresh}{a numeric, threshold to apply. If \code{NA}, defaults to 0.05
for methods that generate a p-value, 0.9 for posterior probabilities from
EBcoexpress and 0.05 on the absolute score from DiffCoEx}
}
\value{
an igraph object, representing the differential network. Scores are
  added as edge attributes with the name 'score'
}
\description{
Threshold the results from a differential co-expression analysis
  and create a differential network.
}
\examples{
#create data
x <- matrix(rnorm(120), 4, 30)
cond <- rep(1:2, 15)

#perform analysis
zscores <- dcScore(x, cond)
pvals <- dcTest(zscores, emat = x, condition = cond)
pvals <- dcAdjust(pvals, p.adjust, method = 'fdr')
ig <- dcNetwork(zscores, pvals, 0.5)

\dontrun{
igraph::plot.igraph(ig)
}

}
\seealso{
\code{\link{dcScore}}, \code{\link{dcTest}}, \code{\link{dcAdjust}}
}
