% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_methods_generic.R
\docType{methods}
\name{diffScore}
\alias{diffScore}
\alias{diffScore,matrix-method}
\alias{diffScore,data.frame-method}
\alias{diffScore,ExpressionSet-method}
\alias{diffScore,SummarizedExperiment-method}
\alias{diffScore,DGEList-method}
\title{Compute scores from differential association analysis}
\usage{
diffScore(emat, condition, dc.method, ...)

\S4method{diffScore}{matrix}(emat, condition, dc.method = "zscore", ...)

\S4method{diffScore}{data.frame}(emat, condition, dc.method = "zscore",
  ...)

\S4method{diffScore}{ExpressionSet}(emat, condition,
  dc.method = "zscore", ...)

\S4method{diffScore}{SummarizedExperiment}(emat, condition,
  dc.method = "zscore", ...)

\S4method{diffScore}{DGEList}(emat, condition, dc.method = "zscore", ...)
}
\arguments{
\item{emat}{a matrix, data.frame, ExpressionSet, SummarizedExperiment or
DGEList}

\item{condition}{a numeric, (with 1's and 2's representing a binary
condition), a factor with 2 levels or a character representing 2 conditions}

\item{dc.method}{a character, representing the method to use. Use
\code{dcMethods()} to get a list of methods}

\item{...}{see details}
}
\value{
a matrix, of scores/statistics representing differential
  associations; p-values will be returned if FTGI is used and posterior
  probabilities if EBcoexpress is used.
}
\description{
Implementations and wrappers for existing implementations for
  methods inferring differential associations/co-expression. This method
  requires a matrix of expression and a binary condition to compute the
  differential association scores for all pairs of features (genes).
  Applications are not limited to analysis of gene expression data and may be
  used for differential associations in general.
}
\details{
When using data from sequencing experiments, make sure appropriate
  filtering for low counts and data transformation has been performed. Not
  doing so will affect estimation of correlation coefficients which most
  methods rely on.

  Additional method specific parameters can be supplied to the function.
  \code{cor.method} can be set to either 'pearson' (default) or 'spearman' to
  determine the method to use for estimating correlations. These are the two
  measures currently suppoerted in the package. We recommend using the
  'spearman' correlation when dealing with sequencing data.

  The beta parameter in the DiffCoEx method can be specified using
  \code{diffcoex.beta} (defaults to 6). This enable soft thresholding of
  correlations similar to WGCNA.

  EBcoexpress specific parameters include \code{ebcoexpress.seed}
  representing the randomisation seed for the EM algorithm,
  \code{ebcoexpress.useBWMC} (defaults to \code{TRUE}) representing whether
  to use the bi-weight mid-correlation coefficient or not and
  \code{ebcoexpress.plot} which plots the diagnostic plots if set to
  \code{TRUE} (defaults to \code{FALSE}).
}
\examples{
x <- matrix(rnorm(60), 2, 30)
cond <- rep(1:2, 15)
diffScore(x, cond) #defaults to zscore
diffScore(x, cond, dc.method = 'diffcoex')

}
\seealso{
\code{\link{dcMethods}}
}
