% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_accessors.R
\name{getSimData}
\alias{getSimData}
\alias{getConditionNames}
\title{Get data and conditions from a given knock-out (KO)}
\usage{
getSimData(simulation, cond.name = NULL, full = FALSE)

getConditionNames(simulation)
}
\arguments{
\item{simulation}{a list, storing data and results generated from simulations}

\item{cond.name}{a character, indicating the knock-out to use to derive
conditions. Multiple knock-outs (KOs) are performed per simulation. If
\code{NULL}, the first KO is chosen}

\item{full}{a logical, indicating whether genes associated with the condition
should be excluded. Defaults to \code{FALSE} and is recommended}
}
\value{
a list, containing \code{emat}, a matrix representing the expression
  data and \code{condition}, a numeric containing the classification of
  samples
}
\description{
Retrieves the simulated expression matrix and sample
  classification for a specific knock-out experiment.
}
\details{
Genes discarded when \code{full} is \code{FALSE} are those that are
  solely dependent on the condition. These genes are discarded from the
  analysis to focus on those that are differentially co-expressed, not
  co-ordinately co-expressed.

  The names of all genes knocked-out can be retrieved using
  \code{getConditionNames}
}
\section{Functions}{
\itemize{
\item \code{getSimData}: get the expression matrix and sample
classification

\item \code{getConditionNames}: get names of the conditions (KOs)
}}

\examples{
data(sim102)
KOs <- getConditionNames(sim102)
simdata <- getSimulatedData(sim102, KOs[2])
cond <- simdata$condition
emat <- simdata$emat
zscores <- dcScore(emat, cond)

}
\seealso{
\code{\link{dcScore}}
}
