% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_methods.R
\name{mi.ap}
\alias{mi.ap}
\title{Mutual information using adaptive partitioning}
\usage{
mi.ap(mat)
}
\arguments{
\item{mat}{a numeric matrix}
}
\value{
matrix of pairwise mutual information estimates
}
\description{
Computes the mutual information between all pairs of variables
  in the matrix (along the columns). Variables are discretised using the
  adaptive partitioning algorithm
}
\examples{
x <- matrix(rnorm(200), 100, 2)
mi.ap(x)
}
