% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_metrics.R
\name{perfMethods}
\alias{perfMethods}
\title{Get names of performance metric methods}
\usage{
perfMethods()
}
\value{
names of methods implemented
}
\description{
Returns a list of performance metrics
}
\examples{
perfMethods()
}
